Rename and get bot logic working
