+++ src/math-games/connect-four/ConnectFourRoot.js
@@ -1,12 +1,12 @@
 import React from 'react'
 
 import { ConnectFourContextProvider } from "./ConnectFourContext";
-import ConnectFourGame from './ConnectFourGame';
+import ConnectFourController from './ConnectFourController';
 
 export default function ConnectFourRoot() {
   return (
     <ConnectFourContextProvider >
-      <ConnectFourGame />
+      <ConnectFourController />
     </ConnectFourContextProvider>
   )
 }
